fix: Rails routes automatically made available as helpers (#322)
* fix: Rails routes automatically made available as helpers

A rails initializer loads the rails helpers.
Rails helper now includes the default routes along with a properly implemented default_url_options method.

* Move prepare_context to a Rails concern

* fix: prevent plugins from being coalesced as into nested arrays

* fix: allow presenter_plugin method to be called multiple times
